LayoutTests/imported/w3c/web-platform-tests/websockets/Secure-Send-binary-65K-arraybuffer.any.worker-expected.txt


FAIL Send 65K binary data on a Secure WebSocket - ArrayBuffer - Connection should be opened assert_equals: expected 0 but got 65000
PASS Send 65K binary data on a Secure WebSocket - ArrayBuffer - Message should be received
FAIL Send 65K binary data on a Secure WebSocket - ArrayBuffer - Connection should be closed assert_true: WebSocket connection should be open expected true got false
